requirements.txt
requests
pillow
transformers
django
transformers
pytorch
pymongo
PyJWT